Update Tased when damaged in game.py
Properly detects if damage has been taken and if health has gone up and prints the Current health.

With the help from Axel we solved a problem were it would not print the health when it went up by moving the line that sets the value of Currenthealth to the bottom after the other two functions had run.