Taser Function
Added a function that uses the module Arduino to be able to control an Arduino using python.

The function gives the digital pin the value 1 which is ON then reads the value of the pin to confirm.

Then after 2 seconds writes 0 to the pin which is OFF then reads the pin to confirm again.